Refactored Simulations -> simulations, dofile handling
The interactive control file is now a property. The contents can be set
in interactive_control_contents string in the testbench, or the filepath
to an existing do-file can be set in interactive_controlfile. If the
contents are given, the file is automatically generated under the
simulation directory.

Simulations is now lowercase, and the subdirectories are named after the
simulation model.