Refine end mechanisms
Ad end_condition attribute. End_condition can be defined as a custom
code snippet determining the simulation end. End_condition is needed for
VHDL as the simulation stops only when all processes have been exited,
and misccmd attribute shoudl be ampty by default.

For verilog this is not necessary, and misccmd could be but the mechanism is created in order
to maintain consistency. for verilog, end_condition is practically an
additional misccmd.