run_rtl: return simulator stdout and stderr outputs
The simulator output logs are currently printed to the terminal, but not
stored anywhere. In A-Core simulations, UART prints are output over
STDERR stream. In order to post-process the values printed over serial,
there needs to be a way to access the simulation logs from python.

This commit changes run_rtl to return the stdout and stderr streams, so
that they are accessible in rtl testbenches.